% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinpicker.R
\name{addin_manager}
\alias{addin_manager}
\title{Addin manager}
\usage{
addin_manager()
}
\description{
An RStudio addin for managing addins! When the addin is lauched, you can select
which addins to install or remove.
}

